# Make test_e2e_{asr,st,mt}_{transformer,conformer}.py faster.