--- a/test/test_e2e_asr_transformer.py
+++ b/test/test_e2e_asr_transformer.py
@@ -1,4 +1,3 @@
-# 18.28s in master, 27162  25010     8%
 import argparse
 import chainer
 import logging