Related #2459. 

- pytest test_e2e_asr_transformer.py: 18 secs -> 5 secs in local
- pytest test_e2e_asr_conformer.py: 17 secs -> 5 secs in local
- pytest test_e2e_st_transformer.py: 25 secs -> 8 secs in local
- pytest test_e2e_mt_transformer.py: 10 secs -> 5 secs in local

Total: (pytest ./e2e_xxx_xxxformer.py) 59 secs -> 12 secs in local

DONE:
- Remove redundant tests (copied from test_e2e_asr_transformer.py), and never-implemented chainer branches
- Replace importlib with real import
- Decrease model and data sizes
- Replace plt.savefig with dummy func